aster_console

Trait AnyConsoleDevice

pub trait AnyConsoleDevice:
    Send
    + Sync
    + Any
    + Debug {
    // Required methods
    fn send(&self, buf: &[u8]);
    fn register_callback(&self, callback: &'static ConsoleCallback);
}

Required Methods

fn send(&self, buf: &[u8])

Sends data to the console device.

fn register_callback(&self, callback: &'static ConsoleCallback)

Registers a callback that will be invoked when the console device receives data.

The callback may be called in the interrupt context. Therefore, it should never sleep.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".