aster_console

Trait AnyConsoleDevice

pub trait AnyConsoleDevice:
    Send
    + Sync
    + Any
    + Debug {
    // Required methods
    fn send(&self, buf: &[u8]);
    fn register_callback(&self, callback: &'static ConsoleCallback);

    // Provided methods
    fn set_font(&self, _font: BitmapFont) -> Result<(), ConsoleSetFontError> { ... }
    fn set_mode(&self, _mode: ConsoleMode) -> bool { ... }
    fn mode(&self) -> Option<ConsoleMode> { ... }
    fn set_keyboard_mode(&self, _mode: KeyboardMode) -> bool { ... }
    fn keyboard_mode(&self) -> Option<KeyboardMode> { ... }
}

Required Methods

fn send(&self, buf: &[u8])

Sends data to the console device.

fn register_callback(&self, callback: &'static ConsoleCallback)

Registers a callback that will be invoked when the console device receives data.

The callback may be called in the interrupt context. Therefore, it should never sleep.

Provided Methods

fn set_font(&self, _font: BitmapFont) -> Result<(), ConsoleSetFontError>

Sets the font of the console device.

fn set_mode(&self, _mode: ConsoleMode) -> bool

Sets the console mode (text or graphics, see mode::ConsoleMode).

Returns true if the mode was changed, false if the mode is not supported.

fn mode(&self) -> Option<ConsoleMode>

Gets the current console mode.

Returns the current console mode, or None if mode switching is not supported.

fn set_keyboard_mode(&self, _mode: KeyboardMode) -> bool

Sets the keyboard mode (see mode::KeyboardMode).

Returns true if the mode was changed, false if the mode is not supported.

fn keyboard_mode(&self) -> Option<KeyboardMode>

Gets the current keyboard mode.

Returns the current keyboard mode, or None if mode switching is not supported.